ObfuscatorExceptionをインナークラスから単独クラスに変更
